requirements.txt
appdirs==1.4.4
Appium-Python-Client==1.1.0
astroid==2.5.1
attrs==20.3.0
black==20.8b1
click==7.1.2
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pylint==2.7.2
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
PyYAML==5.4.1
regex==2020.11.13
selenium==3.141.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wrapt==1.12.1
